- Creating business exceptions  - creating accountService

+++ src/main/java/com/psm/transaction_service/domain/dto/AccountData.java
@@ -1,9 +1,8 @@
 package com.psm.transaction_service.domain.dto;
 
-import java.math.BigDecimal;
-
 public record AccountData(
-    BigDecimal accountId,
+    Long accountId,
     String documentNumber
 ) {
+
 }
